Refactor visualizers with shared base
